+++ ipsuite/nodes.py
@@ -81,6 +81,8 @@ class _Nodes:
     TemperatureCheck = "ipsuite.analysis.TemperatureCheck"
     FixedSphereConstraint = "ipsuite.calculators.FixedSphereConstraint"
     FixedLayerConstraint = "ipsuite.calculators.FixedLayerConstraint"
+    FixedBondLengthConstraint = "ipsuite.calculators.FixedBondLengthConstraint"
+    HookeanConstraint = "ipsuite.calculators.HookeanConstraint"
     AnalyseDensity = "ipsuite.analysis.AnalyseDensity"
 
     # calculators